requirements.txt
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
httplib2==0.9.2
Markdown==2.6.6
psycopg2==2.6.1
pytz==2016.3
requests==2.9.1
six==1.10.0
